+++ indico/modules/rb/client/js/components/TimeInformation.jsx
@@ -50,14 +50,24 @@ function TimeInformation({
             {mEndDate && !mStartDate.isSame(mEndDate, 'day') ? (
               <Translate>
                 <Param
-                  name="startDate"
-                  wrapper={<strong />}
-                  value={mStartDate.format('(ddd.) L')}
+                  name="startDateDay"
+                  wrapper={<span styleName="booking-times-weekday" />}
+                  value={mStartDate.format('ddd.')}
                 />{' '}
-                <Param name="endDate" wrapper={<strong />} value={mEndDate.format('(ddd.) L')} />
+                <Param name="startDate" wrapper={<strong />} value={mStartDate.format('L')} />
+                {' — '}
+                <Param
+                  name="endDateDay"
+                  wrapper={<span styleName="booking-times-weekday" />}
+                  value={mEndDate.format('ddd.')}
+                />{' '}
+                <Param name="endDate" wrapper={<strong />} value={mEndDate.format('L')} />
               </Translate>
             ) : (
-              <strong>{mStartDate.format('(dddd.) L')}</strong>
+              <strong>
+                <span styleName="booking-times-weekday">{mStartDate.format('ddd.')}</span>{' '}
+                {mStartDate.format('L')}
+              </strong>
             )}
           </div>
         </Segment>